copy_backward

+++ main.cpp
@@ -9,6 +9,6 @@ int main()
     v1[0] = 111;
     v1[1] = 222;
     for (int i=11;i<=123;i++) v2.push_back(0.123*i);
-    for (int i=0;i<v2.size();i++) cout<<v2[i]<<endl;
+    for (int i=0;i<(int)v2.size();i++) cout<<v2[i]<<endl;
     return 0;
 }
